# docs/swagger.yaml
definitions:
  controllers.Result:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      message:
        example: 请求成功
        type: string
    type: object
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/zhaozuowu/go-gin-example/blob/master/LICENSE
  termsOfService: https://github.com/zhaozuowu/go-gin-example
  title: Golang Gin API
  version: "1.0"
paths:
  /users/:
    get:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: Page
        in: query
        name: page
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Result'
      summary: 用户列表
      tags:
      - 用户中心
    post:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Result'
      summary: 用户注册接口
      tags:
      - 用户中心
  /users/:id:
    get:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Result'
      summary: 获取用户信息
      tags:
      - 用户中心
swagger: "2.0"